backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
sqlalchemy
pymysql
python-dotenv
cryptography
pydantic[email]
pyjwt
pwdlib[argon2]
python-multipart